refactor(cli): move schedule renderers to schedule_render.go (PRD #1009 M2)
